Use Tokenfield to break up the input (fixes #95)
The csv_to_array function breaks on commas *and* spaces, but the
Tokenfield-based UI is set up so that tags may contain spaces.
Fortunately Tokenfield includes a function to break the field into an
array for you, so I just used that function instead.

(This patch has not been tested.)